requirements.txt
flask
flask_cors
kafka-python
langchain
langchain_community
langchain_openai
python-dotenv
psycopg2-binary
pybreaker
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
PyPDF2
pypdf
schedule
pgvector
PyJWT
redis
apscheduler
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
